+++ bank-feeds/src/codat_bankfeeds/models/shared/bankfeedaccountmapping.py
@@ -84,7 +84,7 @@ class BankFeedAccountMapping(BaseModel):
 
         for n, f in type(self).model_fields.items():
             k = f.alias or n
-            val = serialized.get(k)
+            val = serialized.get(k, serialized.get(n))
             is_nullable_and_explicitly_set = (
                 k in nullable_fields
                 and (self.__pydantic_fields_set__.intersection({n}))  # pylint: disable=no-member
@@ -99,3 +99,9 @@ class BankFeedAccountMapping(BaseModel):
                     m[k] = val
 
         return m
+
+
+try:
+    BankFeedAccountMapping.model_rebuild()
+except NameError:
+    pass