+++ bank-feeds/src/codat_bankfeeds/models/shared/connection.py
@@ -203,7 +203,7 @@ class Connection(BaseModel):
 
         for n, f in type(self).model_fields.items():
             k = f.alias or n
-            val = serialized.get(k)
+            val = serialized.get(k, serialized.get(n))
             is_nullable_and_explicitly_set = (
                 k in nullable_fields
                 and (self.__pydantic_fields_set__.intersection({n}))  # pylint: disable=no-member
@@ -218,3 +218,9 @@ class Connection(BaseModel):
                     m[k] = val
 
         return m
+
+
+try:
+    Connection.model_rebuild()
+except NameError:
+    pass