+++ bank-feeds/src/codat_bankfeeds/models/shared/sourceaccountbatcherrorresponse.py
@@ -30,7 +30,7 @@ class Result(BaseModel):
 
         for n, f in type(self).model_fields.items():
             k = f.alias or n
-            val = serialized.get(k)
+            val = serialized.get(k, serialized.get(n))
 
             if val != UNSET_SENTINEL:
                 if val is not None or k not in optional_fields:
@@ -65,10 +65,20 @@ class SourceAccountBatchErrorResponse(BaseModel):
 
         for n, f in type(self).model_fields.items():
             k = f.alias or n
-            val = serialized.get(k)
+            val = serialized.get(k, serialized.get(n))
 
             if val != UNSET_SENTINEL:
                 if val is not None or k not in optional_fields:
                     m[k] = val
 
         return m
+
+
+try:
+    Result.model_rebuild()
+except NameError:
+    pass
+try:
+    SourceAccountBatchErrorResponse.model_rebuild()
+except NameError:
+    pass