+++ bank-feeds/src/codat_bankfeeds/models/shared/sourceaccountv2.py
@@ -199,7 +199,7 @@ class SourceAccountV2(BaseModel):
 
         for n, f in type(self).model_fields.items():
             k = f.alias or n
-            val = serialized.get(k)
+            val = serialized.get(k, serialized.get(n))
             is_nullable_and_explicitly_set = (
                 k in nullable_fields
                 and (self.__pydantic_fields_set__.intersection({n}))  # pylint: disable=no-member
@@ -214,3 +214,9 @@ class SourceAccountV2(BaseModel):
                     m[k] = val
 
         return m
+
+
+try:
+    SourceAccountV2.model_rebuild()
+except NameError:
+    pass